refactor: controller.js에서 게임 진행부분을 playCarRacing.js로 추출
